LayoutTests/http/tests/loading/resourceLoadStatistics/request-and-deny-storage-access-cross-origin-sandboxed-iframe-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
frame "theIframe" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "theIframe" - didCommitLoadForFrame
frame "theIframe" - didFinishDocumentLoadForFrame
CONFIRM: Do you want to use your  ID on ?
CONFIRM: Do you want to use your  ID on ?
CONSOLE MESSAGE: line 70: PASS document.hasStorageAccess was denied.
frame "theIframe" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "theIframe" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
Tests that cross-origin iframe storage access is denied if the iframe is sandboxed, has the allow token, but the user opts out.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
